refactor: simplify http emitter
