--- src/transport/http/index.ts
@@ -1,2 +1,0 @@
-export * from "./binary_emitter";
-export * from "./structured_emitter";
